requirements.txt
attr==0.3.2
brotli==1.1.0
brotlicffi==1.1.0.0
chardet==5.2.0
cohere
ConfigParser==7.2.0
cryptography==46.0.3
docutils
filelock==3.20.0
h2==4.3.0
HTMLParser==0.0.2
ipython==9.2.0
ipywidgets==8.1.7
js==1.0
keyring==25.6.0
kubernetes
langchain
langchain_chroma
langchain_community
langchain_core
langchain_google_genai
langchain_huggingface
langchain_text_splitters
langchain_together
langchain-postgres
pandas
protobuf==6.33.0
pydantic
pymongo
pyOpenSSL==25.3.0
pytest
python-dotenv
PyMuPDF
rerankers
sentence_transformers==5.1.1
simplejson==3.20.2
Sphinx
SQLAlchemy
streamlit
thread==2.0.5
together
typing_extensions==4.15.0
urllib3_secure_extra==0.1.0
zstandard==0.23.0
psycopg2-binary


langchain-google-genai
langchain-community
langchain-core
sentence-transformers
pathlib
langchain-text-splitters
alembic
pgvector